+++ src/lib/content/cadence-challenges/first-challenge-example/en/overview.ts
@@ -1,14 +1,16 @@
 import type { CadenceChallenge } from '$lib/types/content/cadence-challenge.interface';
 import { ContentTypeEnum } from '$lib/types/content/metadata/content-types.enum';
+import { SubjectsEnum } from '$lib/types/content/metadata/subject.enum';
 
 export const overview: CadenceChallenge = {
 	title: 'First Challenge example',
 	contentType: ContentTypeEnum.Challenge,
 	prize: 10,
 	link: 'https://www.google.com',
-	description: 'This is a challenge example',
+	description:
+		'This is the first challenge description. The first challenge description. This is the first challenge description',
 	image:
-		'https://cdn.discordapp.com/attachments/1054775421671055390/1105958725711319201/tsnakejake_A_cartoon_man_reading_a_mystical_book_with_an_emeral_d5f03067-6692-4152-8ade-37621c0776b5.png',
+		'https://media.discordapp.net/attachments/1054775421671055390/1105958730799005876/Mateo_Roldos_An_avatar_for_the_Emerald_City_student._Emerald_Ci_0d9c4a87-8861-4779-bd9a-080ef0c838a5.png?width=662&height=662',
 	solved: {
 		answerLink: 'https://www.google.com',
 		author: {
@@ -17,5 +19,6 @@ export const overview: CadenceChallenge = {
 			isVerified: false
 		},
 		answerDescription: 'https://www.google.com'
-	}
+	},
+	subjects: [SubjectsEnum.Cadence, SubjectsEnum.Backend]
 };
